requirements.txt
requests~=2.28.1
urllib3~=1.26.13
PyYAML~=6.0
PyGithub~=1.58.0
redis~=4.5.1
requests-cache~=0.9.8